Bug 1309481 - Remote API is missing ability to retrieve and revert historic plug-in and resource configuration
Exposed to remote API the following methods of ConfigurationManager.

findPluginConfigurationUpdatesByCriteria
findResourceConfigurationUpdatesByCriteria
rollbackPluginConfiguration
rollbackResourceConfiguration
