+++ modules/enterprise/server/jar/src/main/java/org/rhq/enterprise/server/configuration/ConfigurationManagerRemote.java
@@ -33,6 +33,8 @@ import org.rhq.core.domain.configuration.group.GroupPluginConfigurationUpdate;
 import org.rhq.core.domain.configuration.group.GroupResourceConfigurationUpdate;
 import org.rhq.core.domain.criteria.GroupPluginConfigurationUpdateCriteria;
 import org.rhq.core.domain.criteria.GroupResourceConfigurationUpdateCriteria;
+import org.rhq.core.domain.criteria.PluginConfigurationUpdateCriteria;
+import org.rhq.core.domain.criteria.ResourceConfigurationUpdateCriteria;
 import org.rhq.core.domain.util.PageList;
 import org.rhq.enterprise.server.authz.PermissionException;
 import org.rhq.enterprise.server.resource.ResourceNotFoundException;
@@ -342,4 +344,47 @@ public interface ConfigurationManagerRemote {
      * @since 4.14
      */
     void purgePluginConfigurationUpdates(Subject subject, int[] configurationUpdateIds, boolean purgeInProgress);
+
+    /**
+     * This method is called when a user has requested to change the resource configuration for an existing resource to
+     * one of its previous values. If the user does not have the proper permissions to change the resource's
+     * configuration, an exception is thrown.
+     *
+     * <p>This will not wait for the agent to finish the configuration update. This will return after the request is
+     * sent. Once the agent finishes with the request, it will send the completed request information to
+     * {@link #completeResourceConfigurationUpdate}.</p>
+     *
+     * @param  subject          the user who is requesting the update
+     * @param  resourceId      identifies the resource to be updated
+     * @param  configHistoryId the id of the resource's previous configuration to rollback to
+     *
+     * @throws ConfigurationUpdateException if the configHistoryId does not exist
+     */
+    void rollbackResourceConfiguration(Subject subject, int resourceId, int configHistoryId)
+        throws ConfigurationUpdateException;
+
+    /**
+     * This method is called when a user has requested to change the plugin configuration for an existing resource to
+     * one of its previous values. If the user does not have the proper permissions to change the resource's
+     * plugin configuration, an exception is thrown.
+     *
+     * <p>This will not wait for the agent to finish the configuration update. This will return after the request is
+     * sent. Once the agent finishes with the request, it will send the completed request information to
+     * {@link #completePluginConfigurationUpdate}.</p>
+     *
+     * @param  subject          the user who is requesting the update
+     * @param  resourceId      identifies the resource to be updated
+     * @param  configHistoryId the id of the resource's previous configuration to rollback to
+     *
+     * @throws ConfigurationUpdateException if the configHistoryId does not exist
+     */
+    void rollbackPluginConfiguration(Subject subject, int resourceId, int configHistoryId)
+        throws ConfigurationUpdateException;
+
+    PageList<ResourceConfigurationUpdate> findResourceConfigurationUpdatesByCriteria(Subject subject,
+        ResourceConfigurationUpdateCriteria criteria);
+
+    PageList<PluginConfigurationUpdate> findPluginConfigurationUpdatesByCriteria(Subject subject,
+        PluginConfigurationUpdateCriteria criteria);
+
 }
